Extract `ComparisonPicker` component
